requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.12.0
pydantic-settings==2.6.1
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
opencv-python-headless==4.10.0.84
tensorflow==2.17.1
streamlit>=1.24
Pillow>=10.3
